backend/requirements.txt
anyio==3.6.2
black==23.1.0
certifi==2022.12.7
click==8.1.3
deta==1.1.0
fastapi==0.92.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pydantic==1.10.5
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
